requirements.txt
torchvision==0.13.0
jsonargparse[signatures]
wandb==0.12.16
gradio>=3.6
pyyaml==5.4.0
protobuf<4.21.0  # 4.21 breaks with wandb, tensorboard, or pytorch-lightning: https://github.com/protocolbuffers/protobuf/issues/10048
lightning_hpo
lightning
gdown
Pillow>=2.9.0
numpy>=1.9.2
tensorflow
diffusers==0.4.1 
tensorboard 
transformers
DeepSpeed 